Move some qt related things into dedicated folder
